eliminate base environment and add abastract environment
